requirements.txt
azure-identity>=1.22.0
boto3~=1.36.4
botocore~=1.36.3
certifi
charset-normalizer
databricks-sql-connector~=4.0.0
idna
jmespath
psycopg2-binary~=2.9.10
python-dateutil
requests~=2.32.3
s3transfer
six
urllib3
pandas~=2.2.3
pydantic~=2.10.5
fsspec
s3fs
pyarrow~=19.0.0
fastparquet
snowflake-connector-python
pyodbc~=5.2.0
cryptography~=44.0.0
PyJWT~=2.10.1
pyrfc6266~=1.0.2
azure-storage-blob>=12.25.1
azure-cli>=2.72.0
azure-core~=1.34.0
